fix: ensure that @types/react is bundled in @previewjs/plugin-react
This should ensure that projects that don't include @types/react in their own dependencies can still have components detected.

While @types/react was correctly declared as a dependency in the preview/ folder, it was not published in the NPM package because the node_modules directory is treated specially (i.e. ignored).

See #138 for an example where this was an issue.