fix(react): support forwardRef() components
This was broken because, when previewing CSF 3 stories, we're looking for a `render` method on the story (as per https://storybook.js.org/blog/component-story-format-3-0).

A component generated with `forwardRef()` has such a method, but it's not meant to be used as directly as a component.

The solution for now is to ignore the `render` property when we're not dealing with a Storybook story module. This does mean that forwardRef components will still be broken if defined in a Storybook file, but that seems like an acceptable tradeoff for now.

A better approach would be to use our knowledge of whether or not the component is a Storybook story to switch rendering logic. This is a little harder because that information isn't currently available in the preview renderer, it needs to be passed through a few layers.